requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.1
SQLAlchemy==2.0.32
gunicorn==21.2.0
python-dotenv==1.0.0